+++ Scenes/CombatManager.tscn
@@ -1,0 +1,6 @@
+[gd_scene load_steps=2 format=3 uid="uid://pwpbnlj2tr13"]
+
+[ext_resource type="Script" path="res://Scripts/CombatManager.cs" id="1_ejcir"]
+
+[node name="CombatManager" type="Node3D"]
+script = ExtResource("1_ejcir")
