+++ Scenes/Player.tscn
@@ -7,7 +7,7 @@
 [sub_resource type="CapsuleShape3D" id="CapsuleShape3D_yd0kh"]
 radius = 0.3
 
-[node name="Player" type="CharacterBody3D" node_paths=PackedStringArray("animPlayer")]
+[node name="Player" type="CharacterBody3D" node_paths=PackedStringArray("animPlayer") groups=["Combatant"]]
 script = ExtResource("1_jfkhu")
 animPlayer = NodePath("body/AnimationPlayer")
