+++ Scripts/CharacterBase.cs
@@ -15,25 +15,12 @@ public abstract partial class CharacterBase : CharacterBody3D
     [Export]
     public int health = 100;
 
-    private Node root;
-
     public override void _Ready()
     {
-        // Register with the TestEnvironment
-        root = GetTree().Root.GetChild(0);
-        if (root is ICombatScene combatScene)
-        {
-            combatScene.RegisterCharacter(this);
-        }
     }
 
     public override void _ExitTree()
     {
-        // Unregister with the TestEnvironment
-        if (root is ICombatScene combatScene)
-        {
-            combatScene.UnregisterCharacter(this);
-        }
     }
 
     public void ReduceHealth(int amount)
